server/requirements.txt
numpy
python-socketio
pyepics
eventlet